Refactor rcs into package
