+++ rcs/__init__.py
@@ -1,0 +1,6 @@
+"""Recursive Consciousness System package."""
+
+from .agent import LangChainRecursiveAgent
+from .experiment import RecursiveConsciousnessExperiment
+
+__all__ = ["LangChainRecursiveAgent", "RecursiveConsciousnessExperiment"]
